Move FunctionalEither to internal package
